Fixed bugs.
Firstly, the regex used by default for JVM services didn't have an end
anchor. This means that a service called, for example, GG_STUBS would
count as a service called GG because the regex used to check for it
would match GG against GG_STUBS.
The second bug was that psutil cmdline() sometimes returns entire
strings for command lines, which need splitting.